third_party/blink/web_tests/virtual/pna-iframes-warning/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=grandparent-expected.txt

This is a testharness.js-based test.
PASS local to local, grandparent navigates: no preflight required.
PASS local to local (same-origin), grandparent navigates: no preflight required.
FAIL public to local, grandparent navigates: failure. assert_equals: expected "timeout" but got "loaded"
PASS public to local, grandparent navigates: success.
Harness: the test ran to completion.
